Code clean up, and change of burn of implemtation and the removal of memberadd.
