+++ eosdactoken/tests/create_accounts.rb
@@ -1,5 +1,5 @@
 require_relative './accounts_helper'
 
-create_accounts(30) # Need to only run this line once.
+create_accounts(10) # Need to only run this line once.
 
 # add_accounts_as_members(30) # The most I could add in one transaction locally was around 250.
